payment + cancel booking

+++ src/main/java/vti/group10/football_booking/repository/PaymentRepository.java
@@ -1,9 +1,7 @@
 package vti.group10.football_booking.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import vti.group10.football_booking.model.Payment;
 
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
-    
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 }
